Add cobb-douglas axis labels - PMT #113565
Also: I realized why django-rest-framework may be complaining about null
values that I've already defined a default for -- In the POST request,
the value explicity gets set to `null`, via the giant object in
GraphMapping.js. I've made a change here to not pass in cobb-douglas
config options when it's not necessary, and I'll continue to refine
this.