NonLinear: fix intersection behavior - PMT #113999
I'm now using a functiongraph for both curves on nonlinear graph - this is
a workaround, see:
https://stackoverflow.com/questions/48773386/jsxgraph-intersection-between-functiongraph-and-line-disappears
This workaround forces me to deal with the curves in a more consistent
way, though. And now positioning is more streamlined for both curves on
this graph. There's now plenty of duplicated code in Graph.js, and can
be refactored, but for now I'm just keeping it in this state without
touching the other graphs too much, since it works.

* Removed unnecessary plotUtils stuff