Assessment system improvements
Add more evaluation tests and action types. Right now I'm deriving the
action type from the field name. This should be sufficient for our
current scope - if necessary we can always add this as a new column in
the spreadsheet model.

I'm also beginning to think about how to integrate this with the student
graph viewer.