Include labels in SVG export - ECON-252
This updates the "Export as SVG" button to use this new `SVGRenderer.dumpToSVG()`
function I made that's currently only available on this dev branch of
jsxgraph:
https://github.com/ccnmtl/jsxgraph/commit/b00d646be472af799ff59ff4cf71413604531d5e#diff-11a1374c93121e44ead6113ba4251ad2R1388

`dumpToSVG()` is basically just the first half of the `dumpToCanvas()` method which
just returns the svg as soon as labels are added, and
all the canvas exporting stuff removed. (dumpToCanvas proved to be
difficult to work with for a few reasons, first:
https://github.com/jsxgraph/jsxgraph/issues/283, and also, I ran into
CORS errors as well as other problems just calling toDataURL() on this
canvas to turn it into a PNG).

The key code is here, which grabs the labels from the DOM and appends it to jsxgraph's svgroot:
```
if (ignoreTexts !== true) {
    // Handle SVG texts
    // Insert all value attributes back into the svg string
    len = values.length;
    for (i = 0; i < len; i++) {
	svg = svg.replace('id="' + values[i][0] + '"', 'id="' + values[i][0] + '" value="' + values[i][1] +'"');
    }
}
```