Don't render a JXGBoard in GraphViewer when gType is null
I've added a simple loading message here.

A JXGBoard in the context of GraphViewer with no gType means the graph
hasn't loaded yet, so we shouldn't be rendering it as it can cause
errors.

From GraphEditor, the gType is declared explicitly with code, rather
than loaded from the database, so this should continue to work fine.

In any case, I'm updating gType to be a required argument of the
JXGBoard component, as we can't render the board if we don't know what
type it is.